Rename class package to classes.
