+++ apps/vor/pkg/classes/store.go
@@ -1,4 +1,4 @@
-package class
+package classes
 
 import (
 	"time"
